+++ pages/index.vue
@@ -1,297 +1,356 @@
 <template>
-  <div class="" id="app">
-    <form v-on:submit.prevent="procesar()" name="form" id="form">
-      <h1>Basic user CRUD</h1>
-      <br />
-      <div class="form-group row col-sm-10">
-        <label class="col-sm-1"> Name </label>
-        <input
-          class="col-sm-5 form-control"
-          placeholder="Name"
-          v-model="name"
-        />
-      </div>
-      <div class="row form-group messaje">
-        <div v-if="submited && !$v.name.required">
-          *The name field is required
-        </div>
-        <div v-if="submited && !$v.name.minLength">
-          *Must have a minimum of 1 characters
-        </div>
-        <div v-if="submited && !$v.name.maxLength">
-          *Must have a maximum of 50 characters
-        </div>
-      </div>
+<div class="wrapper fadeInDown">
+  <div id="formContent">
+    <!-- Tabs Titles -->
 
-      <div class="form-group row col-sm-10">
-        <label class="col-sm-1"> Last Name </label>
-        <input
-          class="form-control col-sm-5"
-          placeholder="Last Name"
-          v-model="lastName"
-        />
-      </div>
-      <div class="row form-group messaje">
-        <div v-if="submited && !$v.lastName.required">
-          *The last name field is required
-        </div>
-        <div v-if="submited && !$v.lastName.minLength">
-          *Must have a minimum of 1 characters
-        </div>
-        <div v-if="submited && !$v.lastName.maxLength">
-          *Must have a maximum of 50 characters
-        </div>
-      </div>
+    <!-- Icon -->
+    <!-- <div class="fadeIn first">
+      <img src="http://danielzawadzki.com/codepen/01/icon.svg" id="icon" alt="User Icon" />
+    </div> -->
 
-      <div class="form-group row col-sm-10">
-        <label class="col-sm-1">Telephone</label>
-        <input
-          type="number"
-          class="form-control col-sm-5"
-          placeholder="Telephone"
-          v-model="telephone"
-        />
-      </div>
-      <div class="row form-group messaje">
-        <div v-if="submited && !$v.telephone.required">
-          *The last telephone field is required
-        </div>
-        <div v-if="submited && !$v.telephone.minLength">
-          *Must have a minimum of 10 characters
-        </div>
-        <div v-if="submited && !$v.telephone.maxLength">
-          *Must have a maximum of 10 charactersss prueba
-        </div>
-      </div>
-
-      <div class="form-group row col-sm-10">
-        <label class="col-sm-1">Address</label>
-        <input
-          class="form-control col-sm-5"
-          placeholder="Address"
-          v-model="address"
-        />
-      </div>
-      <div class="row form-group messaje">
-        <div v-if="submited && !$v.address.required">
-          *The last address field is required
-        </div>
-        <div v-if="submited && !$v.address.minLength">
-          *Must have a minimum of 5 characters
-        </div>
-        <div v-if="submited && !$v.address.maxLength">
-          *Must have a maximum of 20 characters
-        </div>
-      </div>
-
-      <div class="form-group row col-sm-10">
-        <label class="col-sm-1">E-Mail</label>
-        <input
-          class="form-control col-sm-5"
-          placeholder="E-Mail"
-          v-model="email"
-        />
-      </div>
-      <div class="row form-group messaje">
-        <div v-if="submited && !$v.email.required">
-          *The last email field is required
-        </div>
-        <div v-if="submited && !$v.email.email">
-          *You must write a valid email
-        </div>
-      </div>
-      <button type="submit" class="btn btn-success btn-sm" @click="save">
-        Agregar
-      </button>
-      <hr style="background: white" />
-      <table class="table table-success table-dark table-hover" style="background-color: #262a2d !important; border-radius: 1.25rem">
-        <thead>
-          <tr>
-            <th scope="col">Id</th>
-            <th scope="col">Nombre</th>
-            <th scope="col">Apellido</th>
-            <th scope="col">Telefono</th>
-            <th scope="col">Dirección</th>
-            <th scope="col">E-Mail</th>
-          </tr>
-        </thead>
-        <tbody v-for="item of personas" :key="item.idCliente">
-          <tr >
-            <td>{{ item.idCliente }}</td>
-            <td>{{ item.nombres }}</td>
-            <td>{{ item.apellidos }}</td>
-            <td>{{ item.telefono }}</td>
-            <td>{{ item.direccion }}</td>
-            <td>{{ item.email }}</td>
-          </tr>
-        </tbody>
-      </table>
+    <!-- Login Form -->
+    <form v-on:submit.prevent="login">
+      <input type="text" id="login" class="fadeIn second" name="login" placeholder="Login" v-model="usuario">
+      <input type="password" id="password" class="fadeIn third" name="login" placeholder="Password" v-model="password">
+      <input type="submit" class="fadeIn fourth" value="Log In">
     </form>
+
+    <!-- Remind Passowrd -->
+    <div class="alert alert-danger" role="alert" v-if="error">
+      {{error_msg}}
+    </div>
+
   </div>
+</div>
 </template>
 
 <script>
-import {
-  required,
-  minLength,
-  maxLength,
-  email,
-} from "vuelidate/lib/validators";
-import axios from "axios";
-import CrudService from "../pages/services/crudService";
+import axios from 'axios'
 export default {
-  head() {
-    return {
-      title: "Posts App | Home",
-      meta: [
-        {
-          hid: "description",
-          name: "description",
-          content: "Website abut tutorials",
-        },
-      ],
-      htmlAttrs: {
-        lang: "en",
-      },
-    };
+  name: 'Login',
+  layout: 'auth',
+  components:{
+
   },
-  data() {
+  data: function(){
     return {
-      submited: false,
-      name: "",
-      lastName: "",
-      telephone: "",
-      address: "",
-      email: "",
-      user: [],
-      personas: null,
-      persona: {
-        nombres: null,
-        apellidos: null,
-        telefono: null,
-        direccion: null,
-        email: null,
-      },
-    };
-  },
-  methods: {
-    procesar() {
-      /* this.submited = true;
-      if (this.$v.$invalid) {
-        return false;
-      }
-      try {
-        this.user.push({
-          name: this.name,
-          lastName: this.lastName,
-          telephone: this.telephone,
-          address: this.address,
-          email: this.email,
-        }),
-          (this.submited = false);
-        (this.name = ""),
-          (this.lastName = ""),
-          (this.telephone = ""),
-          (this.address = ""),
-          (this.email = "");
-        console.log(this.user);
-        await axios
-          .post("localhost:8080/clientes", this.user)
-          .then((res) => {
-            if (res.status == 200) {
-              console.log(res);
-            }
-          })
-          .catch( (error) => {
-          console.log(error);
-        } );
-      } catch (error) {
-        console.log(error);
-      } */
-    },
-    save() {
-      this.submited = true;
-      if (this.$v.$invalid) {
-        return false;
-      }
-      const persona = {
-        nombres: this.name,
-        apellidos: this.lastName,
-        telefono: this.telephone,
-        direccion: this.address,
-        email: this.email,
-      };
-      console.log(persona);
-      this.crudService.save(persona).then((data) => {
-        if (data.status === 200) {
-          this.getAll();
-          this.submited = false;
-          (this.name = ""),
-          (this.lastName = ""),
-          (this.telephone = ""),
-          (this.address = ""),
-          (this.email = "");
-        }
-        console.log(data);
-      });
-    },
-    select(event){
-      console.log(event);
+      usuario: "",
+      password: "",
+      error: false,
+      error_msg: "",
     }
   },
-  validations: {
-    name: {
-      required,
-      minLength: minLength(1),
-      maxLength: maxLength(50),
-    },
-    lastName: {
-      required,
-      minLength: minLength(1),
-      maxLength: maxLength(50),
-    },
-    telephone: {
-      required,
-      minLength: minLength(9),
-      maxLength: maxLength(10),
-    },
-    address: {
-      required,
-      minLength: minLength(5),
-      maxLength: maxLength(20),
-    },
-    email: {
-      required,
-      email,
-    },
-  },
-  created() {
-    this.crudService = new CrudService();
-  },
-  mounted() {
-    this.crudService.getAll().then((data) => {
-      this.personas = data.data;
-      console.log(this.personas);
-    });
-  },
-};
+  methods: {
+    login (){
+      let json = {
+        "usuario" : this.usuario,
+        "password" : this.password
+      };
+      axios.post('https://api.solodata.es/auth', json)
+      .then( data=>{
+        if(data.data.status == "ok"){
+          localStorage.token = data.data.result.token;
+          this.$router.push('home');
+        }else{
+          this.error = true;
+          this.error_msg = data.data.result.error_msg;
+        }
+      })
+    }
+  }
+
+}
 </script>
-<style>
-.messaje {
-  padding-left: 6rem !important;
-  color: dimgray;
+
+<style scoped>
+html {
+  background-color: #56baed;
 }
 
-.input-validate {
-  border: 1px solid darkred;
-  border-radius: 0.25rem;
-  transition: border-color 0.15s cubic-bezier(0.42, 0, 0.76, 1.1),
-    box-shadow 0.15s ease-in-out;
+body {
+  font-family: "Poppins", sans-serif;
+  height: 100vh;
 }
-.input-validate:hover {
-  border: 1px solid darkred;
-  border-radius: 0.25rem;
-  transition: border-color 0.15s cubic-bezier(0.42, 0, 0.76, 1.1),
-    box-shadow 0.15s ease-in-out;
+
+a {
+  color: #92badd;
+  display:inline-block;
+  text-decoration: none;
+  font-weight: 400;
+}
+
+h2 {
+  text-align: center;
+  font-size: 16px;
+  font-weight: 600;
+  text-transform: uppercase;
+  display:inline-block;
+  margin: 40px 8px 10px 8px;
+  color: #cccccc;
+}
+
+
+
+/* STRUCTURE */
+
+.wrapper {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  justify-content: center;
+  width: 100%;
+  min-height: 100%;
+  padding: 200px;
+}
+
+#formContent {
+  -webkit-border-radius: 10px 10px 10px 10px;
+  border-radius: 10px 10px 10px 10px;
+  background: #fff;
+  padding: 30px;
+  width: 90%;
+  max-width: 450px;
+  position: relative;
+  padding: 0px;
+  -webkit-box-shadow: 0 30px 60px 0 rgba(0,0,0,0.3);
+  box-shadow: 0 30px 60px 0 rgba(0,0,0,0.3);
+  text-align: center;
+}
+
+#formFooter {
+  background-color: #f6f6f6;
+  border-top: 1px solid #dce8f1;
+  padding: 25px;
+  text-align: center;
+  -webkit-border-radius: 0 0 10px 10px;
+  border-radius: 0 0 10px 10px;
+}
+
+
+
+/* TABS */
+
+h2.inactive {
+  color: #cccccc;
+}
+
+h2.active {
+  color: #0d0d0d;
+  border-bottom: 2px solid #5fbae9;
+}
+
+
+
+/* FORM TYPOGRAPHY*/
+
+input[type=button], input[type=submit], input[type=reset]  {
+  background-color: #56baed;
+  border: none;
+  color: white;
+  padding: 15px 80px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  text-transform: uppercase;
+  font-size: 13px;
+  -webkit-box-shadow: 0 10px 30px 0 rgba(95,186,233,0.4);
+  box-shadow: 0 10px 30px 0 rgba(95,186,233,0.4);
+  -webkit-border-radius: 5px 5px 5px 5px;
+  border-radius: 5px 5px 5px 5px;
+  margin: 5px 20px 40px 20px;
+  -webkit-transition: all 0.3s ease-in-out;
+  -moz-transition: all 0.3s ease-in-out;
+  -ms-transition: all 0.3s ease-in-out;
+  -o-transition: all 0.3s ease-in-out;
+  transition: all 0.3s ease-in-out;
+}
+
+input[type=button]:hover, input[type=submit]:hover, input[type=reset]:hover  {
+  background-color: #39ace7;
+}
+
+input[type=button]:active, input[type=submit]:active, input[type=reset]:active  {
+  -moz-transform: scale(0.95);
+  -webkit-transform: scale(0.95);
+  -o-transform: scale(0.95);
+  -ms-transform: scale(0.95);
+  transform: scale(0.95);
+}
+
+input[type=text] {
+  background-color: #f6f6f6;
+  border: none;
+  color: #0d0d0d;
+  padding: 15px 32px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+  margin: 5px;
+  width: 85%;
+  border: 2px solid #f6f6f6;
+  -webkit-transition: all 0.5s ease-in-out;
+  -moz-transition: all 0.5s ease-in-out;
+  -ms-transition: all 0.5s ease-in-out;
+  -o-transition: all 0.5s ease-in-out;
+  transition: all 0.5s ease-in-out;
+  -webkit-border-radius: 5px 5px 5px 5px;
+  border-radius: 5px 5px 5px 5px;
+
+}
+
+input[type=text]:focus {
+  background-color: #fff;
+  border-bottom: 2px solid #5fbae9;
+}
+
+input[type=text]:placeholder {
+  color: #cccccc;
+}
+
+
+input[type=password] {
+  background-color: #f6f6f6;
+  border: none;
+  color: #0d0d0d;
+  padding: 15px 32px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
+  font-size: 16px;
+  margin: 5px;
+  width: 85%;
+  border: 2px solid #f6f6f6;
+  -webkit-transition: all 0.5s ease-in-out;
+  -moz-transition: all 0.5s ease-in-out;
+  -ms-transition: all 0.5s ease-in-out;
+  -o-transition: all 0.5s ease-in-out;
+  transition: all 0.5s ease-in-out;
+  -webkit-border-radius: 5px 5px 5px 5px;
+  border-radius: 5px 5px 5px 5px;
+}
+
+input[type=password]:focus {
+  background-color: #fff;
+  border-bottom: 2px solid #5fbae9;
+}
+
+input[type=password]:placeholder {
+  color: #cccccc;
+}
+
+
+/* ANIMATIONS */
+
+/* Simple CSS3 Fade-in-down Animation */
+.fadeInDown {
+  -webkit-animation-name: fadeInDown;
+  animation-name: fadeInDown;
+  -webkit-animation-duration: 1s;
+  animation-duration: 1s;
+  -webkit-animation-fill-mode: both;
+  animation-fill-mode: both;
+}
+
+@-webkit-keyframes fadeInDown {
+  0% {
+    opacity: 0;
+    -webkit-transform: translate3d(0, -100%, 0);
+    transform: translate3d(0, -100%, 0);
+  }
+  100% {
+    opacity: 1;
+    -webkit-transform: none;
+    transform: none;
+  }
+}
+
+@keyframes fadeInDown {
+  0% {
+    opacity: 0;
+    -webkit-transform: translate3d(0, -100%, 0);
+    transform: translate3d(0, -100%, 0);
+  }
+  100% {
+    opacity: 1;
+    -webkit-transform: none;
+    transform: none;
+  }
+}
+
+/* Simple CSS3 Fade-in Animation */
+@-webkit-keyframes fadeIn { from { opacity:0; } to { opacity:1; } }
+@-moz-keyframes fadeIn { from { opacity:0; } to { opacity:1; } }
+@keyframes fadeIn { from { opacity:0; } to { opacity:1; } }
+
+.fadeIn {
+  opacity:0;
+  -webkit-animation:fadeIn ease-in 1;
+  -moz-animation:fadeIn ease-in 1;
+  animation:fadeIn ease-in 1;
+
+  -webkit-animation-fill-mode:forwards;
+  -moz-animation-fill-mode:forwards;
+  animation-fill-mode:forwards;
+
+  -webkit-animation-duration:1s;
+  -moz-animation-duration:1s;
+  animation-duration:1s;
+}
+
+.fadeIn.first {
+  -webkit-animation-delay: 0.4s;
+  -moz-animation-delay: 0.4s;
+  animation-delay: 0.4s;
+}
+
+.fadeIn.second {
+  -webkit-animation-delay: 0.6s;
+  -moz-animation-delay: 0.6s;
+  animation-delay: 0.6s;
+}
+
+.fadeIn.third {
+  -webkit-animation-delay: 0.8s;
+  -moz-animation-delay: 0.8s;
+  animation-delay: 0.8s;
+}
+
+.fadeIn.fourth {
+  -webkit-animation-delay: 1s;
+  -moz-animation-delay: 1s;
+  animation-delay: 1s;
+}
+
+/* Simple CSS3 Fade-in Animation */
+.underlineHover:after {
+  display: block;
+  left: 0;
+  bottom: -10px;
+  width: 0;
+  height: 2px;
+  background-color: #e3ed56;
+  content: "";
+  transition: width 0.2s;
+}
+
+.underlineHover:hover {
+  color: #0d0d0d;
+}
+
+.underlineHover:hover:after{
+  width: 100%;
+}
+
+
+
+/* OTHERS */
+
+*:focus {
+    outline: none;
+}
+
+#icon {
+  width:60%;
 }
 </style>
